planet's div created for Solar System map Test import image and extern link
